move caveat related code to caveat index file to then re-export from restricted index file
